LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-form-get-expected.txt


FAIL <form> submission with GET method fires navigate event but with formData null assert_equals: expected "replace" but got "push"
